[Xtend2 tests] added support for tests written in Xtend2

+++ tests/org.eclipse.xtext.xtend2.tests/src/org/eclipse/xtext/xtend2/tests/compiler/TestClass1Test.java
@@ -4,8 +4,8 @@ import junit.framework.TestCase;
 
 public class TestClass1Test extends TestCase {
 	public void testFunction() throws Exception {
-		assertEquals(42,new TestClass1().testFunction1().intValue());
-		assertEquals(42,new TestClass1().testFunction2().intValue());
+		assertEquals(42,new TestClass1().testFunction1());
+		assertEquals(42,new TestClass1().testFunction2());
 		assertEquals(42,new TestClass1().testFunction3().intValue());
 		assertEquals(42,new TestClass1().testFunction4().intValue());
 	}
